// types.ts
export enum Section {
  HERO = 'hero',
  EXPLODED = 'exploded',
  TECH_DEEP_DIVE = 'tech-deep-dive',
  COMPARISON = 'comparison',
  MANUFACTURING = 'manufacturing',
  APPLICATIONS = 'applications',
  PARTNERS = 'partners',
  ABOUT_US = 'about-us',
  SUSTAINABILITY = 'sustainability',
  FOOTER = 'footer'
}